separated mode to /commands
